chore: rename pages and reorg components
