+++ src/components/header.tsx
@@ -14,9 +14,9 @@ const headerNavPages = ["home"] as const
 
 type Page =
   | typeof headerNavPages[number]
-  | "conversion-flow"
-  | "send-bitcoin"
-  | "receive-bitcoin"
+  | "convert"
+  | "send"
+  | "receive"
   | "contacts"
   | "transactions"
   | "settings"
